IP.swift: Rework Windows IPv6.bytes
Since there is already special handling for Linux or macOS in IPv6.bytes
it seems more sensible to change that rather than to spoof the linux API

This passes DNSTests.IPTests/testIPv6Bytes on 64-bit Windows 10 21H2.

It also has the nice feature of splitting the work in this branch to the
previous commit which fixed the exclusions from my copy of Foundation,
and this one which I don't expect to be fixed upstream.